# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Field Inspection & Infrastructure Auditor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InspectionIssueType = typeof InspectionIssueType[keyof typeof InspectionIssueType];


export const InspectionIssueType = {
  Pavement_Distress: 'Pavement Distress',
  Drainage: 'Drainage',
  Structural_Damage: 'Structural Damage',
  Signage: 'Signage',
  Erosion: 'Erosion',
  Bridge_Deterioration: 'Bridge Deterioration',
  Utility_Failure: 'Utility Failure',
  Other: 'Other',
} as const;

export type InspectionSeverity = typeof InspectionSeverity[keyof typeof InspectionSeverity];


export const InspectionSeverity = {
  Critical: 'Critical',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type InspectionStatus = typeof InspectionStatus[keyof typeof InspectionStatus];


export const InspectionStatus = {
  Active: 'Active',
  Resolved: 'Resolved',
  Under_Review: 'Under Review',
} as const;

export interface Inspection {
  id: number;
  title: string;
  issueType: InspectionIssueType;
  severity: InspectionSeverity;
  description: string;
  latitude: number;
  longitude: number;
  status: InspectionStatus;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  userId?: string | null;
}

export type InspectionInputIssueType = typeof InspectionInputIssueType[keyof typeof InspectionInputIssueType];


export const InspectionInputIssueType = {
  Pavement_Distress: 'Pavement Distress',
  Drainage: 'Drainage',
  Structural_Damage: 'Structural Damage',
  Signage: 'Signage',
  Erosion: 'Erosion',
  Bridge_Deterioration: 'Bridge Deterioration',
  Utility_Failure: 'Utility Failure',
  Other: 'Other',
} as const;

export type InspectionInputSeverity = typeof InspectionInputSeverity[keyof typeof InspectionInputSeverity];


export const InspectionInputSeverity = {
  Critical: 'Critical',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type InspectionInputStatus = typeof InspectionInputStatus[keyof typeof InspectionInputStatus];


export const InspectionInputStatus = {
  Active: 'Active',
  Resolved: 'Resolved',
  Under_Review: 'Under Review',
} as const;

export interface InspectionInput {
  /** @minLength 1 */
  title: string;
  issueType: InspectionInputIssueType;
  severity: InspectionInputSeverity;
  description: string;
  latitude: number;
  longitude: number;
  status?: InspectionInputStatus;
}

export type InspectionUpdateIssueType = typeof InspectionUpdateIssueType[keyof typeof InspectionUpdateIssueType];


export const InspectionUpdateIssueType = {
  Pavement_Distress: 'Pavement Distress',
  Drainage: 'Drainage',
  Structural_Damage: 'Structural Damage',
  Signage: 'Signage',
  Erosion: 'Erosion',
  Bridge_Deterioration: 'Bridge Deterioration',
  Utility_Failure: 'Utility Failure',
  Other: 'Other',
} as const;

export type InspectionUpdateSeverity = typeof InspectionUpdateSeverity[keyof typeof InspectionUpdateSeverity];


export const InspectionUpdateSeverity = {
  Critical: 'Critical',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type InspectionUpdateStatus = typeof InspectionUpdateStatus[keyof typeof InspectionUpdateStatus];


export const InspectionUpdateStatus = {
  Active: 'Active',
  Resolved: 'Resolved',
  Under_Review: 'Under Review',
} as const;

export interface InspectionUpdate {
  title?: string;
  issueType?: InspectionUpdateIssueType;
  severity?: InspectionUpdateSeverity;
  description?: string;
  latitude?: number;
  longitude?: number;
  status?: InspectionUpdateStatus;
}

export interface InspectionBatchInput {
  inspections: InspectionInput[];
}

export interface DashboardSummary {
  totalLogs: number;
  activeIssues: number;
  resolvedIssues: number;
  regionalHealthScore: number;
  criticalCount: number;
  mediumCount: number;
  lowCount: number;
}

export interface TypeBreakdown {
  issueType: string;
  count: number;
}

export interface SeverityBreakdown {
  severity: string;
  count: number;
}

export interface AppNotification {
  id: number;
  userId: string;
  inspectionId: number;
  inspectionTitle: string;
  message: string;
  type: string;
  read: boolean;
  createdAt: string;
}

export interface MarkReadResult {
  updated: number;
}

export interface ApiError {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};