# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Intelligent Field Inspection & Infrastructure Auditor API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: inspections
    description: Inspection log operations
  - name: dashboard
    description: Dashboard summary and analytics
  - name: notifications
    description: In-app notification feed
  - name: Auth
    description: Browser and mobile authentication endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /inspections:
    get:
      operationId: listInspections
      tags: [inspections]
      summary: List all inspections
      responses:
        "200":
          description: List of inspections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inspection"
    post:
      operationId: createInspection
      tags: [inspections]
      summary: Create an inspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InspectionInput"
      responses:
        "201":
          description: Created inspection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"

  /inspections/batch:
    post:
      operationId: batchCreateInspections
      tags: [inspections]
      summary: Batch create inspections (offline sync)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InspectionBatchInput"
      responses:
        "201":
          description: Batch created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inspection"

  /inspections/{id}:
    get:
      operationId: getInspection
      tags: [inspections]
      summary: Get a single inspection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inspection record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateInspection
      tags: [inspections]
      summary: Update an inspection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InspectionUpdate"
      responses:
        "200":
          description: Updated inspection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteInspection
      tags: [inspections]
      summary: Delete an inspection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Dashboard summary metrics
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/by-type:
    get:
      operationId: getByType
      tags: [dashboard]
      summary: Issues grouped by type
      responses:
        "200":
          description: Breakdown by issue type
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TypeBreakdown"

  /dashboard/by-severity:
    get:
      operationId: getBySeverity
      tags: [dashboard]
      summary: Issues grouped by severity
      responses:
        "200":
          description: Breakdown by severity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeverityBreakdown"

  /dashboard/recent:
    get:
      operationId: getRecentInspections
      tags: [dashboard]
      summary: Recent inspections feed
      responses:
        "200":
          description: Recent inspections (last 10)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inspection"

  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for the current user
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppNotification"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /notifications/mark-all-read:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkReadResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppNotification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Inspection:
      type: object
      required: [id, title, issueType, severity, description, latitude, longitude, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        issueType:
          type: string
          enum:
            - Pavement Distress
            - Drainage
            - Structural Damage
            - Signage
            - Erosion
            - Bridge Deterioration
            - Utility Failure
            - Other
        severity:
          type: string
          enum: [Critical, Medium, Low]
        description:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [Active, Resolved, Under Review]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
        userId:
          type: ["string", "null"]

    InspectionInput:
      type: object
      required: [title, issueType, severity, description, latitude, longitude]
      properties:
        title:
          type: string
          minLength: 1
        issueType:
          type: string
          enum:
            - Pavement Distress
            - Drainage
            - Structural Damage
            - Signage
            - Erosion
            - Bridge Deterioration
            - Utility Failure
            - Other
        severity:
          type: string
          enum: [Critical, Medium, Low]
        description:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [Active, Resolved, Under Review]

    InspectionUpdate:
      type: object
      properties:
        title:
          type: string
        issueType:
          type: string
          enum:
            - Pavement Distress
            - Drainage
            - Structural Damage
            - Signage
            - Erosion
            - Bridge Deterioration
            - Utility Failure
            - Other
        severity:
          type: string
          enum: [Critical, Medium, Low]
        description:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [Active, Resolved, Under Review]

    InspectionBatchInput:
      type: object
      required: [inspections]
      properties:
        inspections:
          type: array
          items:
            $ref: "#/components/schemas/InspectionInput"

    DashboardSummary:
      type: object
      required: [totalLogs, activeIssues, resolvedIssues, regionalHealthScore, criticalCount, mediumCount, lowCount]
      properties:
        totalLogs:
          type: integer
        activeIssues:
          type: integer
        resolvedIssues:
          type: integer
        regionalHealthScore:
          type: integer
        criticalCount:
          type: integer
        mediumCount:
          type: integer
        lowCount:
          type: integer

    TypeBreakdown:
      type: object
      required: [issueType, count]
      properties:
        issueType:
          type: string
        count:
          type: integer

    SeverityBreakdown:
      type: object
      required: [severity, count]
      properties:
        severity:
          type: string
        count:
          type: integer

    AppNotification:
      type: object
      required: [id, userId, inspectionId, inspectionTitle, message, type, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        inspectionId:
          type: integer
        inspectionTitle:
          type: string
        message:
          type: string
        type:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time

    MarkReadResult:
      type: object
      required: [updated]
      properties:
        updated:
          type: integer

    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string