# lib/api-zod/src/generated/types/appNotification.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Field Inspection & Infrastructure Auditor API
 * OpenAPI spec version: 0.1.0
 */

export interface AppNotification {
  id: number;
  userId: string;
  inspectionId: number;
  inspectionTitle: string;
  message: string;
  type: string;
  read: boolean;
  createdAt: Date;
}