# lib/api-zod/src/generated/types/inspection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intelligent Field Inspection & Infrastructure Auditor API
 * OpenAPI spec version: 0.1.0
 */
import type { InspectionIssueType } from './inspectionIssueType';
import type { InspectionSeverity } from './inspectionSeverity';
import type { InspectionStatus } from './inspectionStatus';

export interface Inspection {
  id: number;
  title: string;
  issueType: InspectionIssueType;
  severity: InspectionSeverity;
  description: string;
  latitude: number;
  longitude: number;
  status: InspectionStatus;
  createdAt: Date;
  /** @nullable */
  updatedAt?: Date | null;
  /** @nullable */
  userId?: string | null;
}